fix(daemon-ipc): repair Windows-only build (max_instances type, PathBuf import)
Windows test-infra CI rejected `d247f11` with two Windows-specific
defects in the new owner-only DACL plumbing. Linux build did not catch
either since the relevant code is `#[cfg(windows)]`-gated.

1. `permission/windows_acl.rs::create_first_pipe_instance_owner_only`
   declared `max_instances: u32`, but tokio 1.52 `ServerOptions::
   max_instances(usize)` expects `usize`. Changed the helper signature
   to `usize` to match upstream and drop the now-unneeded conversion at
   the call site (`single_instance::acquire_windows` already passes the
   literal `255`, which infers as `usize`).
2. `ipc/transport/mod.rs` had `use std::path::PathBuf;` at module scope,
   but the `PathBuf` field only lives inside the `#[cfg(unix)] Unix`
   variant. On Windows the import was unused. Cfg-gated the import to
   `#[cfg(unix)]` to silence the warning without hiding any field.

Linux gates (rust:1.80-slim docker, stable 1.95.0)
- cargo check / clippy / fmt: ok
- cargo test --workspace: 333 passed / 0 failed / 2 ignored (unchanged)
- audit-secret-paths.sh: 11/11 PASS

Issue: #26

+++ crates/shikomi-daemon/src/permission/windows_acl.rs
@@ -135,7 +135,7 @@ impl Drop for OwnerOnlySecurityAttributes {
 #[cfg(windows)]
 pub fn create_first_pipe_instance_owner_only(
     pipe_name: &str,
-    max_instances: u32,
+    max_instances: usize,
 ) -> std::io::Result<tokio::net::windows::named_pipe::NamedPipeServer> {
     use tokio::net::windows::named_pipe::ServerOptions;
